+++ python/tskit/tables.py
@@ -32,6 +32,8 @@ import sys
 import warnings
 from dataclasses import dataclass
 from typing import Any
+from typing import Optional
+from typing import Union
 
 import numpy as np
 
@@ -47,12 +49,12 @@ dataclass_options = {"frozen": True}
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class IndividualTableRow:
+class IndividualTableRow(util.Dataclass):
     __slots__ = ["flags", "location", "parents", "metadata"]
     flags: int
     location: np.ndarray
     parents: np.ndarray
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
     # We need a custom eq for the numpy arrays
     def __eq__(self, other):
@@ -67,29 +69,29 @@ class IndividualTableRow:
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class NodeTableRow:
+class NodeTableRow(util.Dataclass):
     __slots__ = ["flags", "time", "population", "individual", "metadata"]
     flags: int
     time: float
     population: int
     individual: int
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class EdgeTableRow:
+class EdgeTableRow(util.Dataclass):
     __slots__ = ["left", "right", "parent", "child", "metadata"]
     left: float
     right: float
     parent: int
     child: int
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class MigrationTableRow:
+class MigrationTableRow(util.Dataclass):
     __slots__ = ["left", "right", "node", "source", "dest", "time", "metadata"]
     left: float
     right: float
@@ -97,27 +99,27 @@ class MigrationTableRow:
     source: int
     dest: int
     time: float
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class SiteTableRow:
+class SiteTableRow(util.Dataclass):
     __slots__ = ["position", "ancestral_state", "metadata"]
     position: float
     ancestral_state: str
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class MutationTableRow:
+class MutationTableRow(util.Dataclass):
     __slots__ = ["site", "node", "derived_state", "parent", "metadata", "time"]
     site: int
     node: int
     derived_state: str
     parent: int
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
     time: float
 
     # We need a custom eq here as we have unknown times (nans) to check
@@ -140,20 +142,20 @@ class MutationTableRow:
 
 @metadata.lazy_decode
 @dataclass(**dataclass_options)
-class PopulationTableRow:
+class PopulationTableRow(util.Dataclass):
     __slots__ = ["metadata"]
-    metadata: Any
+    metadata: Optional[Union[bytes, dict]]
 
 
 @dataclass(**dataclass_options)
-class ProvenanceTableRow:
+class ProvenanceTableRow(util.Dataclass):
     __slots__ = ["timestamp", "record"]
     timestamp: str
     record: str
 
 
 @dataclass(**dataclass_options)
-class TableCollectionIndexes:
+class TableCollectionIndexes(util.Dataclass):
     edge_insertion_order: np.ndarray = None
     edge_removal_order: np.ndarray = None
 
@@ -288,6 +290,27 @@ class BaseTable:
         if index < 0 or index >= len(self):
             raise IndexError("Index out of bounds")
         return self.row_class(*self.ll_table.get_row(index))
+
+    def append(self, row):
+        """
+        Adds a new row to this table and returns the ID of the new row. Metadata, if
+        specified, will be validated and encoded according to the table's
+        :attr:`metadata_schema<tskit.IndividualTable.metadata_schema>`.
+
+        :param row-like row: An object that has attributes corresponding to the
+            properties of the new row. Both the objects returned from ``table[i]`` and
+            e.g. ``ts.individual(i)`` work for this purpose, along with any other
+            object with the correct attributes.
+        :return: The ID of the newly added node.
+        :rtype: int
+        """
+        return self.add_row(
+            **{
+                column: getattr(row, column)
+                for column in self.column_names
+                if "_offset" not in column
+            }
+        )
 
     def clear(self):
         """
@@ -2238,7 +2261,7 @@ class TableCollection:
         return self.metadata_schema.decode_row(self._ll_tables.metadata)
 
     @metadata.setter
-    def metadata(self, metadata: Any) -> None:
+    def metadata(self, metadata: Optional[Union[bytes, dict]]) -> None:
         self._ll_tables.metadata = self.metadata_schema.validate_and_encode_row(
             metadata
         )