+++ python/CHANGELOG.rst
@@ -9,6 +9,9 @@
 
 **Features**
 
+- Add `Table.append` method for adding rows from classes such as `SiteTableRow` and
+  `Site` (:user:`benjeffery`, :issue:`1111`, :pr:`1254`).
+
 - SVG visualization of a single tree allows all mutations on an edge to be plotted
   via the ``all_edge_mutations`` param (:user:`hyanwong`,:issue:`1253`, :pr:`1258`).